test: move action request_review tests in their own file
